requirements.txt
pandas~=2.3.0
streamlit~=1.45.1
requests~=2.32.3
pymilvus==2.5.3
python-dotenv~=1.0.0
openai
transformers
accelerate
langfuse
torch
langgraph
groq